src/requirements.txt
numpy
pandas
scikit-learn
google-cloud-storage
google-cloud-aiplatform
joblib